+++ src/main/java/lotto/domain/NumberGenerationStrategy.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.vo.LottoNumber;
+
 import java.util.List;
 
 @FunctionalInterface
