+++ src/main/java/lotto/domain/RandomNumberGenerationStrategy.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.vo.LottoNumber;
+
 import static java.util.stream.Collectors.*;
 
 import java.util.Collections;
