+++ src/main/java/lotto/domain/TicketMachine.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.entity.LottoTicket;
+
 public class TicketMachine {
 
     public LottoTicket createLottoTicket(NumberGenerationStrategy strategy) {
