+++ src/main/java/lotto/domain/UserLottoResult.java
@@ -1,5 +1,8 @@
 package lotto.domain;
 
+import lotto.domain.enums.LottoGameRank;
+import lotto.domain.vo.Money;
+
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.List;
