+++ src/main/java/lotto/domain/vo/LottoNumbers.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.vo;
 
 import java.util.List;
 import java.util.Objects;
