+++ src/test/java/lotto/domain/TicketMachineTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
 
+import lotto.domain.entity.LottoTicket;
+import lotto.domain.vo.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
