+++ src/test/java/lotto/domain/UserLottoResultTest.java
@@ -8,6 +8,8 @@ import java.math.RoundingMode;
 import java.util.List;
 import java.util.stream.Stream;
 
+import lotto.domain.enums.LottoGameRank;
+import lotto.domain.vo.Money;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
